Move location of the backing-store definition file.
