+++ opencog/persist/api/BackingStore.h
