+++ opencog/persist/api/CMakeLists.txt
@@ -15,3 +15,8 @@ ADD_GUILE_EXTENSION(SCM_CONFIG persist "opencog-ext-path-persist")
 INSTALL (TARGETS persist EXPORT AtomSpaceTargets
 	DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
 )
+
+INSTALL (FILES
+	BackingStore.h
+   DESTINATION "include/opencog/persist/api"
+)
