+++ opencog/persist/sql/multi-driver/SQLPersistSCM.cc
@@ -24,7 +24,7 @@
 #ifdef HAVE_GUILE
 
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atomspace/BackingStore.h>
+#include <opencog/persist/api/BackingStore.h>
 #include <opencog/guile/SchemePrimitive.h>
 
 #include "SQLAtomStorage.h"
